# smallWrite path for import-roaring