--- a/cluster.go
+++ b/cluster.go
@@ -184,6 +184,7 @@ type cluster struct { // nolint: maligned
 	ReplicaN int
 
 	// Threshold for logging long-running queries
+	// TODO(2.0) move this out of cluster. (why is it here??)
 	longQueryTime time.Duration
 
 	// Maximum number of Set() or Clear() commands per request.