@jaffee 2019-03-09
I don't think this adds any new counting overhead – since we were already iterating through the containers, of the incoming bitmap, calls to c.N() are just `return c.n`, and then we're just adding the result to a total.

The `f.storage.Count()` was already there – I'm just using the result now instead of comparing to 0. It could have been switched to `f.storage.Any()`, but now it can't unfortunately.

We don't have *InPlace funcs other than UnionInPlace right now – if you felt an uncontrollable urge to implement them, I definitely would not stop you.

I *think* this change is more or less a strict improvement on what was there, mod updating a counter which I doubt will show up in anyone's benchmarks. I kind of agree that maybe we could use a different heuristic to decide which bitmap to union into which, but I *do* think we need to make that choice explicitly, because the cost of choosing wrong could be quite a lot of allocation.

Maybe just the size of the roaring data in bytes?