@jaffee 2019-03-09
also probably a reasonable choice - I think we'd have to benchmark on different, common workloads to get a feel for which would actually be better, but either of them are probably better than doing that count.